Upgrade to v1.51b
git-svn-id: https://svn.zsnes.com/rep/zsnes@5313 8cdf4848-9e13-0410-98e6-c9478ff1fb58

+++ branches/1.5x/src/tools/depbuild.cpp
@@ -24,6 +24,7 @@ This program generates dependencies for all C/C++/Assembly files
 #include <iostream>
 #include <string>
 #include <cstdio>
+#include <cstdlib>
 using namespace std;
 
 #include "fileutil.h"
@@ -183,7 +184,7 @@ void dependency_calculate(const char *filename, struct stat& stat_buffer)
   }
 }
 
-int main(size_t argc, const char *const *const argv)
+int main(int argc, const char *const *const argv)
 {
   if (argc < 5)
   {
